Implement drive widget
This widget takes care of scanning the system for available drives
(omitting system drives that can cause major harm to the host if
 written).

For now, the widget scans just a single time, but will later be extended
to auto refresh in an interval to detect new connected devices in real
time.

The interaction of this widget is not tested at the moment, due to
issues simulating key presses on cases where stdin was not attached to
the process (CI servers, like Travis and Appveyor).