Right-trim all rows from horizontal & vertical tables
Currently, blank space is added at the right of each row so that all
rows have an equal length. Consider the following example with explicit
quotes:

	'NAME     JOB                     '
	'John Doe Mobile Software Engineer'
	'Jane Doe Lawyer                  '

Now consider that the terminal where this table is printed is not wide
enough to be able to display the second line completely in one line.

Ideally, the long line will wrap to the line below, like this:

	'NAME     JOB            '
	'John Doe Mobile Software'
	Engineer                 '
	'Jane Doe Lawyer         '

However, notice that spaces are added at the right of each row so that
all have the same length, therefore if one of the rows doesn't fit in
the terminal, neither of them do. This results in whitespace wrapping to
the line below as well, causing something like:

	'NAME     JOB            '
	'       '
	'John Doe Mobile Software'
	Engineer'
	'Jane Doe Lawyer         '
	'       '

Which makes the table layout look completely broken.

The solution is to right-trim all the rows of the table, so only the
longest lines wrap, while the rest remain intact.

Notice that is important to not trim from the left as well, since we
support multiple-line values in the tables, for example:

	'NAME     AGE JOB'
	'John Doe 40  Developer'
	'             Musician'
	'Jane Doe 30  Designer'
	'             Actress'

Trimming from the left as well will not lead to the desired result:

	'NAME     AGE JOB'
	'John Doe 40  Developer'
	'Musician'
	'Jane Doe 30  Designer'
	'Actress'

Fixes: https://github.com/resin-io/resin-cli-visuals/issues/18