Run build tasks in sequence
We set `gulp build` as the `prepublish` script in `package.json`. This
means that when running `npm install`, the task will be ran.

Given that we use stdout interceptors in the Progress widget unit tests,
there are some cases where the tests fail for the following reason:

- All tasks in `gulp build` run parallel, therefore sometimes another
task finishes when `gulp test` is still running, interfering with stdout
mocking.

The solution is to run the tasks in `build` sequentially.